rl-host/pyproject.toml
[project]
name = "restaurant-rl-host"
version = "1.0.0"
description = "AI-powered restaurant host using Reinforcement Learning"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Restaurant RL Host Team"}
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "numpy>=1.24.3",
    "pandas>=2.1.3",
    "python-dateutil>=2.8.2",
    "gymnasium>=0.29.1",
    "stable-baselines3>=2.2.1",
    "torch>=2.1.1",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
]

[project.scripts]
rl-host = "backend.app:main"
rl-train = "backend.training.train:main"
rl-eval = "backend.training.evaluate:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_module"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
